Add initial governance documents
This is an initial first cut at the governance documents for OpenHalon.
Initially, I've chosen to write documents around these ideas:

* Committers
* Contributor Onboarding
* Maintainers
* OpenHalon Project Leader (OPL)
* Release Manager

These are well defined roles in Open Source projects, and we'll want to
have these as a part of OpenHalon from the start.

The governance model will evolve as the project progress, but having
a solid foundation will be key to the early success of OpenHalon as an
emerging Open Source project.

Change-Id: I7eec9aa4808f6517164480c805777530af1ede84
Signed-off-by: Kyle Mestery <mestery@mestery.com>