add text for "affiliates" to the CCLA
Before Marvell signed the CCLA, they asked that text be added to cover
"affiliate companies", because they had many such companies worldwide.
Nissa Strottman of HPE was consulted, and she drafted the additional
text, which was added to the copy of the CCLA that Marvell signed.

It may be useful to suggest to the Apache Software Foundation to add
this text to their reference CCLA, for other projects to use.

Change-Id: I9e2167afbb51ffd9798d7811d49a378ea0a6bede
Signed-off-by: Mark Atwood <mark.atwood@hpe.com>
CC: Nissa Strottman <nissa.strottman@hpe.com>